gcp/projects/cuttlefish-hackathon/hackathon-controller/proxy/requirements.txt
Flask==3.1.3
gunicorn==25.1.0
requests==2.32.5
google-auth>=2.49.0
google-cloud-iam>=2.21.0
